CODE REFACTOR
- Refactored the Metabase server initialization in index.ts to streamline the entry point and enhance modularity by importing the MetabaseServer class.
- Changed the default REQUEST_TIMEOUT_MS in config.ts from 30 seconds to 10 minutes for improved request handling.
